Feat: Ajustes nos Services para começar a trabalhar com os dtos

+++ src/main/java/com/br/barbeariaRest/service/ServicoService.java
@@ -1,34 +1,55 @@
 package com.br.barbeariaRest.service;
 
-
+import com.br.barbeariaRest.dto.mapper.ServicoMapper;
+import com.br.barbeariaRest.dto.request.ServicoRequestDTO;
+import com.br.barbeariaRest.dto.response.ServicoResponseDTO;
 import com.br.barbeariaRest.model.Servico;
 import com.br.barbeariaRest.repository.ServicoRepository;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
+@RequiredArgsConstructor
 public class ServicoService {
 
     private final ServicoRepository servicoRepository;
+    private final ServicoMapper servicoMapper;
 
-    public ServicoService(ServicoRepository servicoRepository) {
-        this.servicoRepository = servicoRepository;
+    public List<ServicoResponseDTO> findAll() {
+        return servicoRepository.findAll()
+                .stream()
+                .map(servicoMapper::toResponseDTO)
+                .toList();
     }
 
-    public Servico salvar(Servico servico) {
-        return servicoRepository.save(servico);
+    public ServicoResponseDTO findById(Long id) {
+        Servico servico = servicoRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Serviço não encontrado"));
+        return servicoMapper.toResponseDTO(servico);
     }
 
-    public Servico buscarPorId(Integer id) {
-        return servicoRepository.findById(id).orElse(null);
+    public ServicoResponseDTO create(ServicoRequestDTO dto) {
+        Servico servico = servicoMapper.toEntity(dto);
+        Servico salvo = servicoRepository.save(servico);
+        return servicoMapper.toResponseDTO(salvo);
+    }
 
+    public ServicoResponseDTO update(Long id, ServicoRequestDTO dto) {
+        Servico servico = servicoRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Serviço não encontrado"));
+
+        servicoMapper.updateEntityFromDTO(dto, servico);
+        Servico atualizado = servicoRepository.save(servico);
+
+        return servicoMapper.toResponseDTO(atualizado);
     }
-    public List<Servico> listarTodos() {
-        return servicoRepository.findAll();
-    }
-    public void excluir(Integer id) {
+
+    public void delete(Long id) {
+        if (!servicoRepository.existsById(id)) {
+            throw new RuntimeException("Serviço não encontrado");
+        }
         servicoRepository.deleteById(id);
     }
-
 }
